requirements.txt
requests
pycryptodome
pypiwin32